requirements.txt
PyQt5~=5.15.7
loguru~=0.6.0
python-dotenv~=0.21.0
pyserial~=3.5
qrcode~=7.3.1
django-qrcode~=0.3
requests~=2.28.1